adicionado fluent validations em algumas classes DTO

+++ Versoes.Core.Domain/DataTransferObjects/ProjetoDto.cs
@@ -1,11 +1,10 @@
 ﻿using Versoes.Entities;
-using Versoes.Entities.Models;
 
 namespace Versoes.Core.Domain.DataTransferObjects
 {
     public class ProjetoDto
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Nome { get; set; }
         public StatusDeCadastro Status { get; set; }
     }
